apps/backend-py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
sqlmodel==0.0.21
python-multipart
PyPDF2
python-docx
